Error suppressor @ added libraries/Bcrypt.php:51
is_readable('/dev/urandom') will generate a PHP warning if it's not accessible to the current user. Causes CodeIgniter to halt output. /dev/urandom will not be readable in situations where an open_basedir restriction is in effect. This is very common on shared hosting. Errors from @fopen() are already suppressed in the same expression on line 52. Suppressing the error on is_readable() allows CodeIgniter to continue and use the fallback random generator correctly. I have also seen this occur in phpass.